Tides is a homework that I did for my iOS class at UW, written circa Spring 2012
Code will run on a 6.1 iPhone simulator or an iPhone 6.

Hit the refresh button at the top right corner to populate the screen
with a list of places in your vicinity that have tide data (public)
available. The right arrow next to each city name provides more
detailed tide information.